+++ src/main/java/com/baidu/hugegraph/driver/TraverserManager.java
@@ -19,8 +19,12 @@
 
 package com.baidu.hugegraph.driver;
 
+import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_CAPACITY;
+import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_DEGREE;
+import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_ELEMENTS_LIMIT;
+import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_PAGE_LIMIT;
+
 import java.util.List;
-import java.util.Map;
 
 import com.baidu.hugegraph.api.traverser.CrosspointsAPI;
 import com.baidu.hugegraph.api.traverser.CustomizedCrosspointsAPI;
@@ -39,6 +43,7 @@ import com.baidu.hugegraph.api.traverser.structure.CrosspointsRequest;
 import com.baidu.hugegraph.api.traverser.structure.CustomizedCrosspoints;
 import com.baidu.hugegraph.api.traverser.structure.CustomizedPaths;
 import com.baidu.hugegraph.api.traverser.structure.PathsRequest;
+import com.baidu.hugegraph.api.traverser.structure.Ranks;
 import com.baidu.hugegraph.client.RestClient;
 import com.baidu.hugegraph.structure.constant.Direction;
 import com.baidu.hugegraph.structure.graph.Edge;
@@ -48,12 +53,6 @@ import com.baidu.hugegraph.structure.graph.Shard;
 import com.baidu.hugegraph.structure.graph.Vertex;
 import com.baidu.hugegraph.structure.graph.Vertices;
 import com.baidu.hugegraph.util.E;
-
-import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_PAGE_LIMIT;
-
-import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_CAPACITY;
-import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_DEGREE;
-import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_ELEMENTS_LIMIT;
 
 public class TraverserManager {
 
@@ -216,13 +215,11 @@ public class TraverserManager {
         return this.customizedCrosspointsAPI.post(request);
     }
 
-    public List<Map<Object, Double>> neighborRank(
-                                     NeighborRankAPI.RankRequest request) {
+    public List<Ranks> neighborRank(NeighborRankAPI.Request request) {
         return this.neighborRankAPI.post(request);
     }
 
-    public Map<Object, Double> personalRank(
-                               PersonalRankAPI.RankRequest request) {
+    public Ranks personalRank(PersonalRankAPI.Request request) {
         return this.personalRankAPI.post(request);
     }
