+++ src/main/java/com/baidu/hugegraph/structure/constant/Traverser.java
@@ -22,6 +22,7 @@ package com.baidu.hugegraph.structure.constant;
 public class Traverser {
 
     public static final long DEFAULT_CAPACITY = 10_000_000L;
+    public static final long DEFAULT_LIMIT = 100L;
     public static final long DEFAULT_ELEMENTS_LIMIT = 10_000_000L;
     public static final long DEFAULT_DEGREE = 10_000L;
     public static final long DEFAULT_CROSSPOINT_LIMIT = 10_000L;
