+++ src/test/java/com/baidu/hugegraph/api/ApiTestSuite.java
@@ -32,10 +32,13 @@ import org.junit.runners.Suite;
     EdgeApiTest.class,
     GremlinApiTest.class,
     VariablesApiTest.class,
-    TraverserApiTest.class,
     TaskApiTest.class,
     JobApiTest.class,
-    RestoreApiTest.class
+    RestoreApiTest.class,
+    TraverserApiTest.class,
+    CustomizedPathsApiTest.class,
+    NeighborRankApiTest.class,
+    PersonalRankApiTest.class,
 })
 public class ApiTestSuite {
 }
