+++ src/test/java/com/baidu/hugegraph/api/TraverserApiTest.java
@@ -19,6 +19,8 @@
 
 package com.baidu.hugegraph.api;
 
+import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_PAGE_LIMIT;
+
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
@@ -52,8 +54,6 @@ import com.baidu.hugegraph.testutil.Assert;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
-
-import static com.baidu.hugegraph.structure.constant.Traverser.DEFAULT_PAGE_LIMIT;
 
 public class TraverserApiTest extends BaseApiTest {
 
@@ -599,291 +599,6 @@ public class TraverserApiTest extends BaseApiTest {
     }
 
     @Test
-    public void testCustomizedPathsSourceLabelProperty() {
-        initEdgesWithWeights();
-
-        Object markoId = getVertexId("person", "name", "marko");
-        Object joshId = getVertexId("person", "name", "josh");
-        Object rippleId = getVertexId("software", "name", "ripple");
-        Object lopId = getVertexId("software", "name", "lop");
-
-        PathsRequest.Builder builder = new PathsRequest.Builder();
-        builder.sources().label("person").property("name", "marko");
-        builder.steps().direction(Direction.OUT).labels("knows1")
-                       .weightBy("weight").degree(-1);
-        builder.steps().direction(Direction.OUT).labels("created1")
-                       .weightBy("weight").degree(-1);
-        builder.sortBy(PathsRequest.SortBy.INCR).withVertex(true)
-               .capacity(-1).limit(-1);
-        PathsRequest request = builder.build();
-
-        CustomizedPaths customizedPaths = customizedPathsAPI.post(request);
-        List<CustomizedPaths.Paths> paths = customizedPaths.paths();
-
-        Assert.assertEquals(2, paths.size());
-
-        List<Object> path1 = ImmutableList.of(markoId, joshId, lopId);
-        List<Object> path2 = ImmutableList.of(markoId, joshId, rippleId);
-        Assert.assertEquals(path1, paths.get(0).objects());
-        Assert.assertEquals(path2, paths.get(1).objects());
-
-        List<Double> weights1 = ImmutableList.of(1.0D, 0.4D);
-        List<Double> weights2 = ImmutableList.of(1.0D, 1.0D);
-        Assert.assertEquals(weights1, paths.get(0).weights());
-        Assert.assertEquals(weights2, paths.get(1).weights());
-
-        Set<?> vertices = customizedPaths.vertices().stream()
-                                         .map(Vertex::id)
-                                         .collect(Collectors.toSet());
-        Assert.assertEquals(4, vertices.size());
-        Set<?> expectedVertices = ImmutableSet.of(markoId, lopId,
-                                                  rippleId, joshId);
-        Assert.assertEquals(expectedVertices, vertices);
-        removeEdgesWithWeights();
-    }
-
-    @Test
-    public void testCustomizedPathsSourceIds() {
-        initEdgesWithWeights();
-
-        Object markoId = getVertexId("person", "name", "marko");
-        Object lopId = getVertexId("software", "name", "lop");
-        Object peterId = getVertexId("person", "name", "peter");
-
-        PathsRequest.Builder builder = new PathsRequest.Builder();
-        builder.sources().ids(markoId, peterId);
-        builder.steps().direction(Direction.OUT).labels("created1")
-                       .weightBy("weight").degree(-1);
-        builder.sortBy(PathsRequest.SortBy.INCR).withVertex(true)
-               .capacity(-1).limit(-1);
-        PathsRequest request = builder.build();
-
-        CustomizedPaths customizedPaths = customizedPathsAPI.post(request);
-        List<CustomizedPaths.Paths> paths = customizedPaths.paths();
-
-        Assert.assertEquals(2, paths.size());
-
-        List<Object> path1 = ImmutableList.of(markoId, lopId);
-        List<Object> path2 = ImmutableList.of(peterId, lopId);
-        List<List<Object>> expectedPaths = ImmutableList.of(path1, path2);
-        Assert.assertTrue(expectedPaths.contains(paths.get(0).objects()));
-        Assert.assertTrue(expectedPaths.contains(paths.get(1).objects()));
-
-        List<Double> weights1 = ImmutableList.of(0.2D);
-        List<Double> weights2 = ImmutableList.of(0.4D);
-        List<List<Double>> expectedWeights = ImmutableList.of(weights1,
-                                                              weights2);
-        Assert.assertTrue(expectedWeights.contains(paths.get(0).weights()));
-        Assert.assertTrue(expectedWeights.contains(paths.get(1).weights()));
-
-        Set<?> vertices = customizedPaths.vertices().stream()
-                                         .map(Vertex::id)
-                                         .collect(Collectors.toSet());
-        Assert.assertEquals(3, vertices.size());
-        Set<?> expectedVertices = ImmutableSet.of(markoId, lopId, peterId);
-        Assert.assertEquals(expectedVertices, vertices);
-        removeEdgesWithWeights();
-    }
-
-    @Test
-    public void testCustomizedPathsSourceLabelPropertyMultiValue() {
-        initEdgesWithWeights();
-
-        Object markoId = getVertexId("person", "name", "marko");
-        Object lopId = getVertexId("software", "name", "lop");
-        Object peterId = getVertexId("person", "name", "peter");
-
-        PathsRequest.Builder builder = new PathsRequest.Builder();
-        List<String> names = ImmutableList.of("marko", "peter");
-        builder.sources().label("person").property("name", names);
-        builder.steps().direction(Direction.OUT).labels("created1")
-                       .weightBy("weight").degree(-1);
-        builder.sortBy(PathsRequest.SortBy.INCR).withVertex(true)
-               .capacity(-1).limit(-1);
-        PathsRequest request = builder.build();
-
-        CustomizedPaths customizedPaths = customizedPathsAPI.post(request);
-        List<CustomizedPaths.Paths> paths = customizedPaths.paths();
-
-        Assert.assertEquals(2, paths.size());
-
-        List<Object> path1 = ImmutableList.of(markoId, lopId);
-        List<Object> path2 = ImmutableList.of(peterId, lopId);
-        List<List<Object>> expectedPaths = ImmutableList.of(path1, path2);
-        Assert.assertTrue(expectedPaths.contains(paths.get(0).objects()));
-        Assert.assertTrue(expectedPaths.contains(paths.get(1).objects()));
-
-        List<Double> weights1 = ImmutableList.of(0.2D);
-        List<Double> weights2 = ImmutableList.of(0.4D);
-        List<List<Double>> expectedWeights = ImmutableList.of(weights1,
-                                                              weights2);
-        Assert.assertTrue(expectedWeights.contains(paths.get(0).weights()));
-        Assert.assertTrue(expectedWeights.contains(paths.get(1).weights()));
-
-        Set<?> vertices = customizedPaths.vertices().stream()
-                                         .map(Vertex::id)
-                                         .collect(Collectors.toSet());
-        Assert.assertEquals(3, vertices.size());
-        Set<?> expectedVertices = ImmutableSet.of(markoId, lopId, peterId);
-        Assert.assertEquals(expectedVertices, vertices);
-        removeEdgesWithWeights();
-    }
-
-    @Test
-    public void testCustomizedPathsWithSample() {
-        initEdgesWithWeights();
-
-        Object markoId = getVertexId("person", "name", "marko");
-        Object joshId = getVertexId("person", "name", "josh");
-        Object rippleId = getVertexId("software", "name", "ripple");
-        Object lopId = getVertexId("software", "name", "lop");
-
-        PathsRequest.Builder builder = new PathsRequest.Builder();
-        builder.sources().label("person").property("name", "marko");
-        builder.steps().direction(Direction.OUT).labels("knows1")
-                       .weightBy("weight").degree(-1);
-        builder.steps().direction(Direction.OUT).labels("created1")
-                       .weightBy("weight").degree(-1).sample(1);
-        builder.sortBy(PathsRequest.SortBy.INCR).withVertex(true)
-               .capacity(-1).limit(-1);
-        PathsRequest request = builder.build();
-
-        CustomizedPaths customizedPaths = customizedPathsAPI.post(request);
-        List<CustomizedPaths.Paths> paths = customizedPaths.paths();
-
-        Assert.assertEquals(1, paths.size());
-
-        List<Object> path1 = ImmutableList.of(markoId, joshId, rippleId);
-        List<Object> path2 = ImmutableList.of(markoId, joshId, lopId);
-        List<List<Object>> expectedPaths = ImmutableList.of(path1, path2);
-        Assert.assertTrue(expectedPaths.contains(paths.get(0).objects()));
-
-        List<Double> weights1 = ImmutableList.of(1D, 0.4D);
-        List<Double> weights2 = ImmutableList.of(1D, 1D);
-
-        Assert.assertTrue(weights1.equals(paths.get(0).weights()) ||
-                          weights2.equals(paths.get(0).weights()));
-
-        Set<?> vertices = customizedPaths.vertices().stream()
-                                         .map(Vertex::id)
-                                         .collect(Collectors.toSet());
-        Assert.assertEquals(3, vertices.size());
-        Assert.assertTrue(path1.containsAll(vertices) ||
-                          path2.containsAll(vertices));
-        removeEdgesWithWeights();
-    }
-
-    @Test
-    public void testCustomizedPathsWithDecr() {
-        initEdgesWithWeights();
-
-        Object markoId = getVertexId("person", "name", "marko");
-        Object joshId = getVertexId("person", "name", "josh");
-        Object rippleId = getVertexId("software", "name", "ripple");
-        Object lopId = getVertexId("software", "name", "lop");
-
-        PathsRequest.Builder builder = new PathsRequest.Builder();
-        builder.sources().label("person").property("name", "marko");
-        builder.steps().direction(Direction.OUT).labels("knows1")
-                       .weightBy("weight").degree(-1);
-        builder.steps().direction(Direction.OUT).labels("created1")
-                       .weightBy("weight").degree(-1);
-        builder.sortBy(PathsRequest.SortBy.DECR).withVertex(true)
-               .capacity(-1).limit(-1);
-        PathsRequest request = builder.build();
-
-        CustomizedPaths customizedPaths = customizedPathsAPI.post(request);
-        List<CustomizedPaths.Paths> paths = customizedPaths.paths();
-
-        Assert.assertEquals(2, paths.size());
-
-        List<Object> path1 = ImmutableList.of(markoId, joshId, rippleId);
-        List<Object> path2 = ImmutableList.of(markoId, joshId, lopId);
-        Assert.assertEquals(path1, paths.get(0).objects());
-        Assert.assertEquals(path2, paths.get(1).objects());
-
-        List<Double> weights1 = ImmutableList.of(1.0D, 1.0D);
-        List<Double> weights2 = ImmutableList.of(1.0D, 0.4D);
-        Assert.assertEquals(weights1, paths.get(0).weights());
-        Assert.assertEquals(weights2, paths.get(1).weights());
-
-        Set<?> vertices = customizedPaths.vertices().stream()
-                                         .map(Vertex::id)
-                                         .collect(Collectors.toSet());
-        Assert.assertEquals(4, vertices.size());
-        Set<?> expectedVertices = ImmutableSet.of(markoId, lopId,
-                                                  rippleId, joshId);
-        Assert.assertEquals(expectedVertices, vertices);
-        removeEdgesWithWeights();
-    }
-
-    @Test
-    public void testCustomizedPathsWithLimit() {
-        initEdgesWithWeights();
-
-        Object markoId = getVertexId("person", "name", "marko");
-        Object joshId = getVertexId("person", "name", "josh");
-        Object rippleId = getVertexId("software", "name", "ripple");
-        Object lopId = getVertexId("software", "name", "lop");
-
-        PathsRequest.Builder builder = new PathsRequest.Builder();
-        builder.sources().label("person").property("name", "marko");
-        builder.steps().direction(Direction.OUT).labels("knows1")
-                       .weightBy("weight").degree(-1);
-        builder.steps().direction(Direction.OUT).labels("created1")
-                       .weightBy("weight").degree(-1);
-        builder.sortBy(PathsRequest.SortBy.INCR).withVertex(true)
-               .capacity(-1).limit(1);
-        PathsRequest request = builder.build();
-
-        CustomizedPaths customizedPaths = customizedPathsAPI.post(request);
-        List<CustomizedPaths.Paths> paths = customizedPaths.paths();
-
-        Assert.assertEquals(1, paths.size());
-
-        List<Object> path1 = ImmutableList.of(markoId, joshId, lopId);
-        List<Object> path2 = ImmutableList.of(markoId, joshId, rippleId);
-        List<List<Object>> expectedPaths = ImmutableList.of(path1, path2);
-        Assert.assertTrue(expectedPaths.contains(paths.get(0).objects()));
-
-        List<Double> weights1 = ImmutableList.of(1.0D, 0.4D);
-        List<Double> weights2 = ImmutableList.of(1.0D, 1.0D);
-        List<List<Double>> expectedWeights = ImmutableList.of(weights1,
-                                                              weights2);
-        Assert.assertTrue(expectedWeights.contains(paths.get(0).weights()));
-
-        Set<?> vertices = customizedPaths.vertices().stream()
-                                         .map(Vertex::id)
-                                         .collect(Collectors.toSet());
-        Assert.assertEquals(3, vertices.size());
-        Set<?> vertices1 = ImmutableSet.of(markoId, lopId, joshId);
-        Set<?> vertices2 = ImmutableSet.of(markoId, lopId, rippleId);
-        Set<Set<?>> expectedVertices = ImmutableSet.of(vertices1, vertices2);
-        Assert.assertTrue(expectedVertices.contains(vertices));
-        removeEdgesWithWeights();
-    }
-
-    @Test
-    public void testCustomizedPathsWithCapacity() {
-        initEdgesWithWeights();
-
-        PathsRequest.Builder builder = new PathsRequest.Builder();
-        builder.sources().label("person").property("name", "marko");
-        builder.steps().direction(Direction.OUT).labels("knows1")
-                       .weightBy("weight").degree(-1);
-        builder.steps().direction(Direction.OUT).labels("created1")
-                       .weightBy("weight").degree(-1);
-        builder.sortBy(PathsRequest.SortBy.INCR).withVertex(true)
-               .capacity(1).limit(-1);
-        PathsRequest request = builder.build();
-
-        Assert.assertThrows(ServerException.class, () -> {
-            customizedPathsAPI.post(request);
-        });
-        removeEdgesWithWeights();
-    }
-
-    @Test
     public void testCustomizedCrosspoints() {
         Object lopId = getVertexId("software", "name", "lop");
         Object joshId = getVertexId("person", "name", "josh");
@@ -915,92 +630,6 @@ public class TraverserApiTest extends BaseApiTest {
                                                .collect(Collectors.toSet());
         List<Object> expectedVids = ImmutableList.of(rippleId, joshId, lopId);
         Assert.assertTrue(expectedVids.containsAll(vertices));
-    }
-
-    @Test
-    public void testNeighborRank() {
-        initNeighborRankGraph();
-
-        NeighborRankAPI.RankRequest.Builder builder;
-        builder = new NeighborRankAPI.RankRequest.Builder();
-        builder.source("O");
-        builder.steps().direction(Direction.OUT).degree(-1).top(10);
-        builder.steps().direction(Direction.OUT).degree(-1).top(10);
-        builder.steps().direction(Direction.OUT).degree(-1).top(10);
-        builder.alpha(0.9).capacity(-1);
-        NeighborRankAPI.RankRequest request = builder.build();
-
-        List<Ranks> ranks = neighborRankAPI.post(request);
-        Assert.assertEquals(4, ranks.size());
-        Assert.assertEquals(ImmutableMap.of("O", 1.0D), ranks.get(0));
-
-        Assert.assertEquals(ImmutableMap.of("B", 0.45075000000000004D,
-                                            "A", 0.3D,
-                                            "C", 0.3D),
-                            ranks.get(1));
-        Assert.assertEquals(ImmutableMap.builder()
-                                        .put("G", 0.17550000000000002D)
-                                        .put("H", 0.17550000000000002D)
-                                        .put("E", 0.135D)
-                                        .put("F", 0.135D)
-                                        .put("I", 0.135D)
-                                        .put("J", 0.135D)
-                                        .build(),
-                            ranks.get(2));
-        Assert.assertEquals(ImmutableMap.of("M", 0.15795D,
-                                            "K", 0.12150000000000001D,
-                                            "L", 0.12150000000000001D),
-                            ranks.get(3));
-    }
-
-    @Test
-    public void testNeighborRankWithTop() {
-        initNeighborRankGraph();
-
-        NeighborRankAPI.RankRequest.Builder builder;
-        builder = new NeighborRankAPI.RankRequest.Builder();
-        builder.source("O");
-        builder.steps().direction(Direction.OUT).degree(-1).top(2);
-        builder.steps().direction(Direction.OUT).degree(-1).top(3);
-        builder.steps().direction(Direction.OUT).degree(-1).top(2);
-        builder.alpha(0.9).capacity(-1);
-        NeighborRankAPI.RankRequest request = builder.build();
-
-        List<Ranks> ranks = neighborRankAPI.post(request);
-        Assert.assertEquals(4, ranks.size());
-        Assert.assertEquals(ImmutableMap.of("O", 1.0D), ranks.get(0));
-
-        Assert.assertEquals(ImmutableMap.of("B", 0.45075000000000004D,
-                                            "A", 0.3D),
-                            ranks.get(1));
-        Assert.assertEquals(ImmutableMap.of("G", 0.17550000000000002D,
-                                            "H", 0.17550000000000002D,
-                                            "E", 0.135D),
-                            ranks.get(2));
-        Assert.assertEquals(ImmutableMap.of("M", 0.15795D,
-                                            "K", 0.12150000000000001D),
-                            ranks.get(3));
-    }
-
-    @Test
-    public void testPersonalRank() {
-        initPersonalRankGraph();
-
-        PersonalRankAPI.RankRequest.Builder builder;
-        builder = new PersonalRankAPI.RankRequest.Builder();
-        builder.source("A").label("like").alpha(0.9).maxDepth(50).build();
-        PersonalRankAPI.RankRequest request = builder.build();
-
-        Map<Object, Double> ranks = personalRankAPI.post(request);
-        Assert.assertEquals(ImmutableMap.builder()
-                                        .put("A", 0.22378692172243492D)
-                                        .put("B", 0.2065750574989044D)
-                                        .put("c", 0.18972188606657317)
-                                        .put("a", 0.1460239032907022)
-                                        .put("C", 0.09839507219265439)
-                                        .put("d", 0.08959757100230095)
-                                        .put("b", 0.04589958822642998)
-                                        .build(), ranks);
     }
 
     @Test
@@ -1147,168 +776,5 @@ public class TraverserApiTest extends BaseApiTest {
         Assert.assertThrows(ServerException.class, () -> {
             edgesAPI.shards(1 * 1024 * 1024 - 1);
         });
-    }
-
-
-    private static void initEdgesWithWeights() {
-        SchemaManager schema = schema();
-        schema.edgeLabel("knows1")
-              .sourceLabel("person")
-              .targetLabel("person")
-              .properties("date", "weight")
-              .nullableKeys("weight")
-              .ifNotExist()
-              .create();
-
-        schema.edgeLabel("created1")
-              .sourceLabel("person")
-              .targetLabel("software")
-              .properties("date", "weight")
-              .nullableKeys("weight")
-              .ifNotExist()
-              .create();
-
-        Vertex marko = getVertex("person", "name", "marko");
-        Vertex vadas = getVertex("person", "name", "vadas");
-        Vertex lop = getVertex("software", "name", "lop");
-        Vertex josh = getVertex("person", "name", "josh");
-        Vertex ripple = getVertex("software", "name", "ripple");
-        Vertex peter = getVertex("person", "name", "peter");
-
-        marko.addEdge("knows1", vadas, "date", "20160110", "weight", 0.5);
-        marko.addEdge("knows1", josh, "date", "20130220", "weight", 1.0);
-        marko.addEdge("created1", lop, "date", "20171210", "weight", 0.4);
-        josh.addEdge("created1", lop, "date", "20091111", "weight", 0.4);
-        josh.addEdge("created1", ripple, "date", "20171210", "weight", 1.0);
-        peter.addEdge("created1", lop, "date", "20170324", "weight", 0.2);
-    }
-
-    private static void removeEdgesWithWeights() {
-        List<Long> elTaskIds = new ArrayList<>();
-        EdgeLabel knows1 = schema().getEdgeLabel("knows1");
-        EdgeLabel created1 = schema().getEdgeLabel("created1");
-        ImmutableList.of(knows1, created1).forEach(edgeLabel -> {
-            elTaskIds.add(edgeLabelAPI.delete(edgeLabel.name()));
-        });
-        elTaskIds.forEach(BaseApiTest::waitUntilTaskCompleted);
-    }
-
-    private static void initNeighborRankGraph() {
-        GraphManager graph = graph();
-        SchemaManager schema = schema();
-
-        schema.vertexLabel("m_person")
-              .properties("name")
-              .useCustomizeStringId()
-              .ifNotExist()
-              .create();
-
-        schema.vertexLabel("movie")
-              .properties("name")
-              .useCustomizeStringId()
-              .ifNotExist()
-              .create();
-
-        schema.edgeLabel("follow")
-              .sourceLabel("m_person")
-              .targetLabel("m_person")
-              .ifNotExist()
-              .create();
-
-        schema.edgeLabel("like")
-              .sourceLabel("m_person")
-              .targetLabel("movie")
-              .ifNotExist()
-              .create();
-
-        schema.edgeLabel("directedBy")
-              .sourceLabel("movie")
-              .targetLabel("m_person")
-              .ifNotExist()
-              .create();
-
-        Vertex O = graph.addVertex(T.label, "m_person", T.id, "O", "name", "O");
-
-        Vertex A = graph.addVertex(T.label, "m_person", T.id, "A", "name", "A");
-        Vertex B = graph.addVertex(T.label, "m_person", T.id, "B", "name", "B");
-        Vertex C = graph.addVertex(T.label, "m_person", T.id, "C", "name", "C");
-        Vertex D = graph.addVertex(T.label, "m_person", T.id, "D", "name", "D");
-
-        Vertex E = graph.addVertex(T.label, "movie", T.id, "E", "name", "E");
-        Vertex F = graph.addVertex(T.label, "movie", T.id, "F", "name", "F");
-        Vertex G = graph.addVertex(T.label, "movie", T.id, "G", "name", "G");
-        Vertex H = graph.addVertex(T.label, "movie", T.id, "H", "name", "H");
-        Vertex I = graph.addVertex(T.label, "movie", T.id, "I", "name", "I");
-        Vertex J = graph.addVertex(T.label, "movie", T.id, "J", "name", "J");
-
-        Vertex K = graph.addVertex(T.label, "m_person", T.id, "K", "name", "K");
-        Vertex L = graph.addVertex(T.label, "m_person", T.id, "L", "name", "L");
-        Vertex M = graph.addVertex(T.label, "m_person", T.id, "M", "name", "M");
-
-        O.addEdge("follow", A);
-        O.addEdge("follow", B);
-        O.addEdge("follow", C);
-        D.addEdge("follow", O);
-
-        A.addEdge("follow", B);
-        A.addEdge("like", E);
-        A.addEdge("like", F);
-
-        B.addEdge("like", G);
-        B.addEdge("like", H);
-
-        C.addEdge("like", I);
-        C.addEdge("like", J);
-
-        E.addEdge("directedBy", K);
-        F.addEdge("directedBy", B);
-        F.addEdge("directedBy", L);
-
-        G.addEdge("directedBy", M);
-    }
-
-    private static void initPersonalRankGraph() {
-        GraphManager graph = graph();
-        SchemaManager schema = schema();
-
-        schema.propertyKey("name").asText().ifNotExist().create();
-
-        schema.vertexLabel("person")
-              .properties("name")
-              .useCustomizeStringId()
-              .ifNotExist()
-              .create();
-
-        schema.vertexLabel("movie")
-              .properties("name")
-              .useCustomizeStringId()
-              .ifNotExist()
-              .create();
-
-        schema.edgeLabel("like")
-              .sourceLabel("person")
-              .targetLabel("movie")
-              .ifNotExist()
-              .create();
-
-        Vertex A = graph.addVertex(T.label, "person", T.id, "A", "name", "A");
-        Vertex B = graph.addVertex(T.label, "person", T.id, "B", "name", "B");
-        Vertex C = graph.addVertex(T.label, "person", T.id, "C", "name", "C");
-
-        Vertex a = graph.addVertex(T.label, "movie", T.id, "a", "name", "a");
-        Vertex b = graph.addVertex(T.label, "movie", T.id, "b", "name", "b");
-        Vertex c = graph.addVertex(T.label, "movie", T.id, "c", "name", "c");
-        Vertex d = graph.addVertex(T.label, "movie", T.id, "d", "name", "d");
-
-        A.addEdge("like", a);
-        A.addEdge("like", c);
-
-        B.addEdge("like", a);
-        B.addEdge("like", b);
-        B.addEdge("like", c);
-        B.addEdge("like", d);
-
-        C.addEdge("like", c);
-        C.addEdge("like", d);
     }
 }
